requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
python-docx>=1.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
PyMuPDF>=1.23.0
pymilvus>=2.4.0
openai>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
nest_asyncio>=1.6.0
